move hashid to utils, move by_name functions to models
remove hash_encoder function
