+++ server/utils.py
@@ -1,15 +1,19 @@
-import server.models as models
+from hashids import Hashids
+from werkzeug.routing import BaseConverter, ValidationError
+
 from server.extensions import cache
 
-# TODO : Better form of caching this.
+class HashidConverter(BaseConverter):
+    # ID hashing configuration.
+    # DO NOT CHANGE ONCE THE APP IS PUBLICLY AVAILABLE. You will break every
+    # link with an ID in it.
+    hashids = Hashids(min_length=6)
 
-def assignment_by_name(name, course_offering=None):
-    """ Return assignment object when given a name. If a course offering is
-    provided, the assignment name is prefixed by the course offering.
-    """
-    if course_offering:
-        name = course_offering + '/' + name
-    return models.Assignment.query.filter_by(name=name).one_or_none()
+    def to_python(self, value):
+        numbers = self.hashids.decode(value)
+        if len(numbers) != 1:
+            raise ValidationError('Could not decode hash {} into ID'.format(value))
+        return numbers[0]
 
-def course_by_name(name):
-    return models.Course.query.filter_by(offering=name).one_or_none()
+    def to_url(self, value):
+        return self.hashids.encode(value)
